requirements.txt
fastapi
uvicorn
python-dotenv
simple-salesforce
groq
pandas
openpyxl
reportlab
jinja2